// src/application/dtos/user/userDetailResponse.dto.ts
export class UserDetailResponseDto {
  id: number = 0;
  address: string = "";
  age: string = "";
  birthday: string = "";
  cellPhone: string = "";
  documentNumber: string = "";
  email: string = "";
  idDocumentsType: number = 0;
  idRole: number = 0;
  lastName: string = "";
  names: string = "";
  state: string = "";
}